fix: move formula to formula_builder file
